feat(write.go): add support for specifying target to write configuration
The `Write` struct now has a new field `Target` which allows specifying the target to write the configuration. The default value is "action".

A new function `WithTarget` is added to the `Write` struct which takes a string parameter `target` and sets it as the value of the `Target` field. If the `target` parameter is empty, it returns an error indicating that the target cannot be empty.